ref(profiling): Remove profile summary page and redirect to transaction summary profiles tab
The standalone profile summary page at /explore/profiling/summary/:projectId/
is no longer needed — the transaction summary's Profiles tab provides
the same functionality. Replace the route with a redirect that preserves
query parameters.

Also remove the differential flamegraph route at
/explore/profiling/profile/:projectId/differential-flamegraph/ which was
only reachable from the profile summary page. Delete all components and
utilities that become unused as a result.
