--- static/app/components/profiling/flamegraph/flamegraphDrawer/differentialFlamegraphDrawer.tsx
@@ -1,418 +1,0 @@
-import type {MouseEventHandler} from 'react';
-import {memo, useCallback, useMemo, useState} from 'react';
-import styled from '@emotion/styled';
-
-import {Button} from '@sentry/scraps/button';
-import {Checkbox} from '@sentry/scraps/checkbox';
-import {Flex} from '@sentry/scraps/layout';
-import {Tooltip} from '@sentry/scraps/tooltip';
-
-import {ExportProfileButton} from 'sentry/components/profiling/exportProfileButton';
-import {IconPanel} from 'sentry/icons';
-import {t} from 'sentry/locale';
-import type {
-  CanvasPoolManager,
-  CanvasScheduler,
-} from 'sentry/utils/profiling/canvasScheduler';
-import {filterFlamegraphTree} from 'sentry/utils/profiling/filterFlamegraphTree';
-import type {Flamegraph} from 'sentry/utils/profiling/flamegraph';
-import type {FlamegraphPreferences} from 'sentry/utils/profiling/flamegraph/flamegraphStateProvider/reducers/flamegraphPreferences';
-import {useFlamegraphPreferences} from 'sentry/utils/profiling/flamegraph/hooks/useFlamegraphPreferences';
-import {useDispatchFlamegraphState} from 'sentry/utils/profiling/flamegraph/hooks/useFlamegraphState';
-import type {FlamegraphFrame} from 'sentry/utils/profiling/flamegraphFrame';
-import type {ProfileGroup} from 'sentry/utils/profiling/profile/importProfile';
-import {invertCallTree} from 'sentry/utils/profiling/profile/utils';
-import {useLocalStorageState} from 'sentry/utils/useLocalStorageState';
-import {useOrganization} from 'sentry/utils/useOrganization';
-import {useParams} from 'sentry/utils/useParams';
-
-import {FlamegraphTreeTable} from './flamegraphTreeTable';
-
-interface DifferentialFlamegraphDrawerProps {
-  canvasPoolManager: CanvasPoolManager;
-  canvasScheduler: CanvasScheduler;
-  flamegraph: Flamegraph;
-  formatDuration: Flamegraph['formatter'];
-  getFrameColor: (frame: FlamegraphFrame) => string;
-  profileGroup: ProfileGroup;
-  referenceNode: FlamegraphFrame;
-  rootNodes: FlamegraphFrame[];
-  onResize?: MouseEventHandler<HTMLElement>;
-  onResizeReset?: MouseEventHandler<HTMLElement>;
-}
-
-const DifferentialFlamegraphDrawer = memo(function FlamegraphDrawer(
-  props: DifferentialFlamegraphDrawerProps
-) {
-  const params = useParams();
-  const orgSlug = useOrganization().slug;
-  const flamegraphPreferences = useFlamegraphPreferences();
-  const dispatch = useDispatchFlamegraphState();
-
-  const [tab, setTab] = useLocalStorageState<'bottom up' | 'top down'>(
-    'profiling-drawer-view',
-    'bottom up'
-  );
-  const [treeType, setTreeType] = useState<'all' | 'application' | 'system'>('all');
-  const [recursion, setRecursion] = useState<'collapsed' | null>(null);
-
-  const maybeFilteredOrInvertedTree: FlamegraphFrame[] | null = useMemo(() => {
-    const skipFunction: (f: FlamegraphFrame) => boolean =
-      treeType === 'application'
-        ? f => !f.frame.is_application
-        : treeType === 'system'
-          ? f => f.frame.is_application
-          : () => false;
-
-    const maybeFilteredRoots =
-      treeType === 'all'
-        ? props.rootNodes
-        : filterFlamegraphTree(props.rootNodes, skipFunction);
-
-    if (tab === 'top down') {
-      return maybeFilteredRoots;
-    }
-
-    return invertCallTree(maybeFilteredRoots);
-  }, [tab, treeType, props.rootNodes]);
-
-  const handleRecursionChange = useCallback(
-    (evt: React.ChangeEvent<HTMLInputElement>) => {
-      setRecursion(evt.currentTarget.checked ? 'collapsed' : null);
-    },
-    []
-  );
-
-  const onBottomUpClick = useCallback(() => {
-    setTab('bottom up');
-  }, [setTab]);
-
-  const onTopDownClick = useCallback(() => {
-    setTab('top down');
-  }, [setTab]);
-
-  const onAllApplicationsClick = useCallback(() => {
-    setTreeType('all');
-  }, []);
-
-  const onApplicationsClick = useCallback(() => {
-    setTreeType('application');
-  }, []);
-
-  const onSystemsClick = useCallback(() => {
-    setTreeType('system');
-  }, []);
-
-  const onTableLeftClick = useCallback(() => {
-    dispatch({type: 'set layout', payload: 'table left'});
-  }, [dispatch]);
-
-  const onTableBottomClick = useCallback(() => {
-    dispatch({type: 'set layout', payload: 'table bottom'});
-  }, [dispatch]);
-
-  const onTableRightClick = useCallback(() => {
-    dispatch({type: 'set layout', payload: 'table right'});
-  }, [dispatch]);
-
-  return (
-    <FrameDrawer layout={flamegraphPreferences.layout}>
-      <ProfilingDetailsFrameTabs>
-        <ProfilingDetailsListItem className={tab === 'bottom up' ? 'active' : undefined}>
-          <Button
-            data-title={t('Bottom Up')}
-            priority="link"
-            size="zero"
-            onClick={onBottomUpClick}
-          >
-            {t('Bottom Up')}
-          </Button>
-        </ProfilingDetailsListItem>
-        <ProfilingDetailsListItem
-          margin="none"
-          className={tab === 'top down' ? 'active' : undefined}
-        >
-          <Button
-            data-title={t('Top Down')}
-            priority="link"
-            size="zero"
-            onClick={onTopDownClick}
-          >
-            {t('Top Down')}
-          </Button>
-        </ProfilingDetailsListItem>
-        <Separator />
-        <ProfilingDetailsListItem className={treeType === 'all' ? 'active' : undefined}>
-          <Button
-            data-title={t('All Frames')}
-            priority="link"
-            size="zero"
-            onClick={onAllApplicationsClick}
-          >
-            {t('All Frames')}
-          </Button>
-        </ProfilingDetailsListItem>
-        <ProfilingDetailsListItem
-          className={treeType === 'application' ? 'active' : undefined}
-        >
-          <Button
-            data-title={t('Application Frames')}
-            priority="link"
-            size="zero"
-            onClick={onApplicationsClick}
-          >
-            {t('Application Frames')}
-          </Button>
-        </ProfilingDetailsListItem>
-        <ProfilingDetailsListItem
-          margin="none"
-          className={treeType === 'system' ? 'active' : undefined}
-        >
-          <Button
-            data-title={t('System Frames')}
-            priority="link"
-            size="zero"
-            onClick={onSystemsClick}
-          >
-            {t('System Frames')}
-          </Button>
-        </ProfilingDetailsListItem>
-        <Separator />
-        <ProfilingDetailsListItem>
-          <FrameDrawerLabel>
-            <Checkbox
-              size="xs"
-              checked={recursion === 'collapsed'}
-              onChange={handleRecursionChange}
-            />
-            {t('Collapse recursion')}
-          </FrameDrawerLabel>
-        </ProfilingDetailsListItem>
-        <ProfilingDetailsListItem
-          style={{
-            flex: '1 1 100%',
-            cursor:
-              flamegraphPreferences.layout === 'table bottom' ? 'ns-resize' : undefined,
-          }}
-          onMouseDown={
-            flamegraphPreferences.layout === 'table bottom' ? props.onResize : undefined
-          }
-          onDoubleClick={
-            flamegraphPreferences.layout === 'table bottom'
-              ? props.onResizeReset
-              : undefined
-          }
-        />
-        <ProfilingDetailsListItem margin="none">
-          <ExportProfileButton
-            size="zero"
-            eventId={params.eventId}
-            projectId={params.projectId}
-            orgId={orgSlug}
-            disabled={params.eventId === undefined || params.projectId === undefined}
-          />
-        </ProfilingDetailsListItem>
-        <Separator />
-        <ProfilingDetailsListItem>
-          <Flex align="center" gap="2xs" height="100%">
-            <Tooltip title={t('Table left')} skipWrapper>
-              <StyledButton
-                // @ts-expect-error transparent is not a valid priority in legacy UI
-                priority="transparent"
-                active={flamegraphPreferences.layout === 'table left'}
-                onClick={onTableLeftClick}
-                title={t('Table left')}
-                aria-label={t('Table left')}
-                size="xs"
-                icon={<IconPanel direction="left" />}
-              />
-            </Tooltip>
-            <Tooltip title={t('Table bottom')} skipWrapper>
-              <StyledButton
-                // @ts-expect-error transparent is not a valid priority in legacy UI
-                priority="transparent"
-                active={flamegraphPreferences.layout === 'table bottom'}
-                onClick={onTableBottomClick}
-                title={t('Table bottom')}
-                aria-label={t('Table bottom')}
-                size="xs"
-                icon={<IconPanel direction="down" />}
-              />
-            </Tooltip>
-            <Tooltip title={t('Table right')} skipWrapper>
-              <StyledButton
-                // @ts-expect-error transparent is not a valid priority in legacy UI
-                priority="transparent"
-                active={flamegraphPreferences.layout === 'table right'}
-                onClick={onTableRightClick}
-                title={t('Table right')}
-                aria-label={t('Table right')}
-                size="xs"
-                icon={<IconPanel direction="right" />}
-              />
-            </Tooltip>
-          </Flex>
-        </ProfilingDetailsListItem>
-      </ProfilingDetailsFrameTabs>
-
-      <FlamegraphTreeTable
-        {...props}
-        expanded={tab === 'top down'}
-        onTopDownClick={onTopDownClick}
-        onBottomUpClick={onBottomUpClick}
-        recursion={recursion}
-        flamegraph={props.flamegraph}
-        referenceNode={props.referenceNode}
-        tree={maybeFilteredOrInvertedTree ?? []}
-        canvasScheduler={props.canvasScheduler}
-        canvasPoolManager={props.canvasPoolManager}
-      />
-
-      {flamegraphPreferences.layout === 'table left' ||
-      flamegraphPreferences.layout === 'table right' ? (
-        <ResizableVerticalDrawer>
-          {/* The border should be 1px, but we want the actual handler to be wider
-          to improve the user experience and not have users have to click on the exact pixel */}
-          <InvisibleHandler
-            onMouseDown={props.onResize}
-            onDoubleClick={props.onResizeReset}
-          />
-        </ResizableVerticalDrawer>
-      ) : null}
-    </FrameDrawer>
-  );
-});
-
-const ResizableVerticalDrawer = styled('div')`
-  width: 1px;
-  grid-area: drawer;
-  background-color: ${p => p.theme.tokens.background.tertiary};
-  position: relative;
-`;
-
-const InvisibleHandler = styled('div')`
-  opacity: 0;
-  width: ${p => p.theme.space.md};
-  position: absolute;
-  inset: 0;
-  cursor: ew-resize;
-  transform: translateX(-50%);
-  background-color: transparent;
-`;
-
-const FrameDrawerLabel = styled('label')`
-  display: flex;
-  align-items: center;
-  white-space: nowrap;
-  margin-bottom: 0;
-  height: 100%;
-  font-weight: ${p => p.theme.font.weight.sans.regular};
-  gap: ${p => p.theme.space.xs};
-`;
-
-// Linter produces a false positive for the grid layout. I did not manage to find out
-// how to "fix it" or why it is not working, I imagine it could be due to the ternary?
-const FrameDrawer = styled('div')<{layout: FlamegraphPreferences['layout']}>`
-  display: grid;
-  grid-template-rows: auto 1fr;
-  grid-template-columns: ${({layout}) =>
-    layout === 'table left' ? '1fr auto' : layout === 'table right' ? 'auto 1fr' : '1fr'};
-  /* false positive for grid layout */
-  /* stylelint-disable */
-  grid-template-areas: ${({layout}) =>
-    layout === 'table bottom'
-      ? `
-    'tabs tabs'
-    'table details'
-    'drawer drawer'
-    `
-      : layout === 'table left'
-        ? `
-      'tabs tabs drawer'
-      'table table drawer'
-      'details details drawer';
-      `
-        : `
-      'drawer tabs tabs'
-      'drawer table table'
-      'drawer details details';
-      `};
-`;
-const Separator = styled('li')`
-  width: 1px;
-  height: 66%;
-  margin: 0 ${p => p.theme.space.md};
-  border: 1px solid ${p => p.theme.tokens.border.primary};
-  transform: translateY(29%);
-`;
-
-const ProfilingDetailsFrameTabs = styled('ul')`
-  display: flex;
-  list-style-type: none;
-  padding: 0 ${p => p.theme.space.md};
-  margin: 0;
-  border-top: 1px solid ${prop => prop.theme.tokens.border.primary};
-  background-color: ${props => props.theme.tokens.background.secondary};
-  user-select: none;
-  grid-area: tabs;
-`;
-
-const ProfilingDetailsListItem = styled('li')<{
-  margin?: 'none';
-  size?: 'sm';
-}>`
-  height: 100%;
-  display: flex;
-  align-items: center;
-  font-size: ${p => p.theme.font.size.sm};
-  margin-right: ${p => (p.margin === 'none' ? 0 : p.theme.space.md)};
-
-  button {
-    height: 100%;
-    border: none;
-    border-top: 2px solid transparent;
-    border-bottom: 2px solid transparent;
-    border-radius: 0;
-    font-weight: ${p => p.theme.font.weight.sans.regular};
-    margin: 0;
-    padding: ${p => (p.size === 'sm' ? p.theme.space['2xs'] : p.theme.space.xs)} 0;
-    color: ${p => p.theme.tokens.content.primary};
-    max-height: ${p => (p.size === 'sm' ? '24px' : undefined)};
-
-    &::after {
-      display: block;
-      content: attr(data-title);
-      font-weight: ${p => p.theme.font.weight.sans.medium};
-      height: 1px;
-      color: transparent;
-      overflow: hidden;
-      visibility: hidden;
-      white-space: nowrap;
-    }
-
-    &:hover {
-      color: ${p => p.theme.tokens.content.primary};
-    }
-  }
-
-  &.active button {
-    font-weight: ${p => p.theme.font.weight.sans.medium};
-    border-bottom: 2px solid ${prop => prop.theme.tokens.border.accent};
-  }
-`;
-
-const StyledButton = styled('button')<{active: boolean}>`
-  padding: ${p => p.theme.space.xs} ${p => p.theme.space.xs};
-  opacity: ${p => (p.active ? 0.7 : 0.5)};
-  background-color: transparent;
-
-  display: flex !important;
-  align-items: center;
-  justify-content: center;
-
-  &:hover {
-    opacity: ${p => (p.active ? 0.6 : 0.5)};
-  }
-`;
-
-export {DifferentialFlamegraphDrawer};
