+++ static/app/utils/profiling/routes.tsx
@@ -37,46 +37,6 @@ function generateContinuousProfileFlamechartRoute({
   return `/organizations/${organization.slug}/${PROFILING_BASE_PATHNAME}/profile/${projectSlug}/flamegraph/`;
 }
 
-function generateProfileDifferentialFlamegraphRoute({
-  organization,
-  projectSlug,
-}: {
-  organization: Organization;
-  projectSlug: Project['slug'];
-}): string {
-  return `/organizations/${organization.slug}/${PROFILING_BASE_PATHNAME}/profile/${projectSlug}/differential-flamegraph/`;
-}
-
-export function generateProfileDifferentialFlamegraphRouteWithQuery({
-  organization,
-  projectSlug,
-  query,
-  fingerprint,
-  transaction,
-  breakpoint,
-}: {
-  breakpoint: number;
-  fingerprint: number;
-  organization: Organization;
-  projectSlug: Project['slug'];
-  transaction: string;
-  query?: Location['query'];
-}): LocationDescriptor {
-  const pathname = generateProfileDifferentialFlamegraphRoute({
-    organization,
-    projectSlug,
-  });
-  return {
-    pathname,
-    query: {
-      ...query,
-      transaction,
-      fingerprint,
-      breakpoint,
-    },
-  };
-}
-
 export function generateProfilingRouteWithQuery({
   organization,
   query,
