--- static/app/views/profiling/profileSummary/slowestProfileFunctions.tsx
@@ -1,309 +1,0 @@
-import {useCallback, useMemo, useState} from 'react';
-import styled from '@emotion/styled';
-
-import type {SelectOption} from '@sentry/scraps/compactSelect';
-import {CompactSelect} from '@sentry/scraps/compactSelect';
-import {Flex} from '@sentry/scraps/layout';
-import {Link} from '@sentry/scraps/link';
-import {OverlayTrigger} from '@sentry/scraps/overlayTrigger';
-
-import {Count} from 'sentry/components/count';
-import {LoadingIndicator} from 'sentry/components/loadingIndicator';
-import {Pagination} from 'sentry/components/pagination';
-import {PerformanceDuration} from 'sentry/components/performanceDuration';
-import {TextTruncateOverflow} from 'sentry/components/profiling/textTruncateOverflow';
-import {t, tn} from 'sentry/locale';
-import type {Organization} from 'sentry/types/organization';
-import type {Project} from 'sentry/types/project';
-import {defined} from 'sentry/utils';
-import {trackAnalytics} from 'sentry/utils/analytics';
-import {Frame} from 'sentry/utils/profiling/frame';
-import type {EventsResultsDataRow} from 'sentry/utils/profiling/hooks/types';
-import {useCurrentProjectFromRouteParam} from 'sentry/utils/profiling/hooks/useCurrentProjectFromRouteParam';
-import {useProfileFunctions} from 'sentry/utils/profiling/hooks/useProfileFunctions';
-import {formatSort} from 'sentry/utils/profiling/hooks/utils';
-import {generateProfileRouteFromProfileReference} from 'sentry/utils/profiling/routes';
-import {decodeScalar} from 'sentry/utils/queryString';
-import {MutableSearch} from 'sentry/utils/tokenizeSearch';
-import {useLocation} from 'sentry/utils/useLocation';
-import {useNavigate} from 'sentry/utils/useNavigate';
-import {useOrganization} from 'sentry/utils/useOrganization';
-
-const SLOWEST_FUNCTIONS_LIMIT = 15;
-const SLOWEST_FUNCTIONS_CURSOR = 'functionsCursor';
-
-const functionsFields = [
-  'package',
-  'function',
-  'count()',
-  'p75()',
-  'sum()',
-  'all_examples()',
-] as const;
-
-type FunctionsField = (typeof functionsFields)[number];
-interface SlowestProfileFunctionsProps {
-  transaction: string;
-}
-
-export function SlowestProfileFunctions(props: SlowestProfileFunctionsProps) {
-  const navigate = useNavigate();
-  const location = useLocation();
-  const organization = useOrganization();
-  const project = useCurrentProjectFromRouteParam();
-  const [functionType, setFunctionType] = useState<'application' | 'system' | 'all'>(
-    'application'
-  );
-  const functionsCursor = useMemo(
-    () => decodeScalar(location.query[SLOWEST_FUNCTIONS_CURSOR]),
-    [location.query]
-  );
-  const functionsSort = useMemo(
-    () =>
-      formatSort<FunctionsField>(
-        decodeScalar(location.query.functionsSort),
-        functionsFields,
-        {
-          key: 'sum()',
-          order: 'desc',
-        }
-      ),
-    [location.query.functionsSort]
-  );
-
-  const handleFunctionsCursor = useCallback(
-    (cursor: any, pathname: any, query: any) =>
-      navigate({
-        pathname,
-        query: {...query, [SLOWEST_FUNCTIONS_CURSOR]: cursor},
-      }),
-    [navigate]
-  );
-
-  const query = useMemo(() => {
-    const conditions = new MutableSearch('');
-    conditions.setFilterValues('transaction', [props.transaction]);
-    if (functionType === 'application') {
-      conditions.setFilterValues('is_application', ['1']);
-    } else if (functionType === 'system') {
-      conditions.setFilterValues('is_application', ['0']);
-    }
-    return conditions.formatString();
-  }, [functionType, props.transaction]);
-
-  const functionsQuery = useProfileFunctions<FunctionsField>({
-    fields: functionsFields,
-    referrer: 'api.profiling.profile-summary-functions-table',
-    sort: functionsSort,
-    query,
-    limit: SLOWEST_FUNCTIONS_LIMIT,
-    cursor: functionsCursor,
-  });
-
-  const onChangeFunctionType = useCallback((v: any) => setFunctionType(v.value), []);
-  const functions = functionsQuery.data?.data ?? [];
-
-  const onSlowestFunctionClick = useCallback(() => {
-    trackAnalytics('profiling_views.go_to_flamegraph', {
-      organization,
-      source: 'unknown',
-    });
-  }, [organization]);
-
-  return (
-    <SlowestFunctionsContainer>
-      <Flex justify="between" align="center" marginBottom="md">
-        <SlowestFunctionsTypeSelect
-          value={functionType}
-          options={SLOWEST_FUNCTION_OPTIONS}
-          onChange={onChangeFunctionType}
-          trigger={triggerProps => (
-            <OverlayTrigger.Button {...triggerProps} priority="transparent" size="zero" />
-          )}
-          offset={4}
-        />
-        <SlowestFunctionsPagination
-          pageLinks={functionsQuery.getResponseHeader?.('Link')}
-          onCursor={handleFunctionsCursor}
-          size="xs"
-        />
-      </Flex>
-      <SlowestFunctionsList>
-        {functionsQuery.isPending ? (
-          <SlowestFunctionsQueryState>
-            <LoadingIndicator size={36} />
-          </SlowestFunctionsQueryState>
-        ) : functionsQuery.isError ? (
-          <SlowestFunctionsQueryState>
-            {t('Failed to fetch slowest functions')}
-          </SlowestFunctionsQueryState>
-        ) : functions.length ? (
-          functions.map((fn, i) => {
-            return (
-              <SlowestFunctionEntry
-                key={i}
-                func={fn}
-                organization={organization}
-                project={project}
-                onSlowestFunctionClick={onSlowestFunctionClick}
-              />
-            );
-          })
-        ) : (
-          <SlowestFunctionsQueryState>
-            {t('The fastest code is one that never runs.')}
-          </SlowestFunctionsQueryState>
-        )}
-      </SlowestFunctionsList>
-    </SlowestFunctionsContainer>
-  );
-}
-
-interface SlowestFunctionEntryProps {
-  func: EventsResultsDataRow<'function' | 'package' | 'count()' | 'p75()' | 'sum()'>;
-  onSlowestFunctionClick: () => void;
-  organization: Organization;
-  project: Project | null;
-}
-function SlowestFunctionEntry(props: SlowestFunctionEntryProps) {
-  const organization = useOrganization();
-  const frame = useMemo(() => {
-    return new Frame(
-      {
-        key: 0,
-        name: props.func.function as string,
-        package: props.func.package as string,
-      },
-      // Ensures that the frame runs through the normalization code path
-      props.project?.platform && /node|javascript/.test(props.project.platform)
-        ? props.project.platform
-        : undefined,
-      'aggregate'
-    );
-  }, [props.func, props.project]);
-
-  let rendered = <TextTruncateOverflow>{frame.name}</TextTruncateOverflow>;
-  // @ts-expect-error TS(7053): Element implicitly has an 'any' type because expre... Remove this comment to see the full error message
-  const example = props.func['all_examples()']?.[0];
-  if (defined(example)) {
-    const target = generateProfileRouteFromProfileReference({
-      organization,
-      projectSlug: props.project?.slug ?? '',
-      frameName: frame.name,
-      framePackage: frame.package as string,
-      reference: example,
-    });
-    rendered = (
-      <Link onClick={props.onSlowestFunctionClick} to={target}>
-        {rendered}
-      </Link>
-    );
-  }
-
-  return (
-    <SlowestFunctionRow>
-      <SlowestFunctionMainRow>
-        <div>{rendered}</div>
-        <div>
-          <PerformanceDuration nanoseconds={props.func['sum()'] as number} abbreviation />
-        </div>
-      </SlowestFunctionMainRow>
-      <SlowestFunctionMetricsRow>
-        <div>
-          <TextTruncateOverflow>{frame.package}</TextTruncateOverflow>
-        </div>
-        <div>
-          <Count value={props.func['count()'] as number} />{' '}
-          {tn('time', 'times', props.func['count()'])}
-          {', '}
-          <PerformanceDuration nanoseconds={props.func['p75()'] as number} abbreviation />
-        </div>
-      </SlowestFunctionMetricsRow>
-    </SlowestFunctionRow>
-  );
-}
-
-const SlowestFunctionsList = styled('div')`
-  flex-basis: 100%;
-  overflow: auto;
-  min-height: 0;
-`;
-
-const SlowestFunctionsContainer = styled('div')`
-  margin-top: ${p => p.theme.space.xs};
-  min-height: 0;
-  display: flex;
-  flex-direction: column;
-  padding: 0 ${p => p.theme.space.md};
-  border-bottom: 1px solid ${p => p.theme.tokens.border.primary};
-`;
-
-const SlowestFunctionsPagination = styled(Pagination)`
-  margin: 0;
-
-  button {
-    height: 16px;
-    width: 16px;
-    min-width: 16px;
-    min-height: 16px;
-
-    svg {
-      width: 10px;
-      height: 10px;
-    }
-  }
-`;
-
-const SlowestFunctionsTypeSelect = styled(CompactSelect)`
-  button {
-    margin: 0;
-    padding: 0;
-  }
-`;
-
-const SlowestFunctionsQueryState = styled('div')`
-  text-align: center;
-  padding: ${p => p.theme.space.xl} ${p => p.theme.space.xs};
-  color: ${p => p.theme.tokens.content.secondary};
-`;
-
-const SlowestFunctionRow = styled('div')`
-  margin-bottom: ${p => p.theme.space.md};
-`;
-
-const SlowestFunctionMainRow = styled('div')`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-
-  > div:first-child {
-    min-width: 0;
-  }
-  > div:last-child {
-    white-space: nowrap;
-  }
-`;
-
-const SlowestFunctionMetricsRow = styled('div')`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  font-size: ${p => p.theme.font.size.sm};
-  color: ${p => p.theme.tokens.content.secondary};
-  margin-top: ${p => p.theme.space['2xs']};
-`;
-
-const SLOWEST_FUNCTION_OPTIONS: Array<SelectOption<'application' | 'system' | 'all'>> = [
-  {
-    label: t('Slowest Application Functions'),
-    value: 'application' as const,
-  },
-  {
-    label: t('Slowest System Functions'),
-    value: 'system' as const,
-  },
-  {
-    label: t('Slowest Functions'),
-    value: 'all' as const,
-  },
-];
